+++ src/App.js
@@ -1,16 +1,16 @@
 import React from "react";
 import {BrowserRouter as Router, Routes, Route} from 'react-router-dom';
-import Home from './Home';
-import About from "./About";
-import Products from "./Products";
-import Contact from "./Contact";
-import SingleProduct from "./SingleProduct";
-import Cart from "./Cart";
-import Error from "./Error";
 import { ThemeProvider } from "styled-components";
 import { GlobalStyle } from "./GlobalStyle";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
+import Home from "./pages/Home";
+import About from "./pages/About";
+import Products from "./pages/Products";
+import Contact from "./pages/Contact";
+import SingleProduct from "./pages/SingleProduct";
+import Cart from "./pages/Cart";
+import Error from "./pages/Error";
 
 const App = () => {
   const theme = {
