+++ src/pages/Home.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import styled from 'styled-components'
-import HeroSection from './components/HeroSection'
-import Services from './components/Services'
-import Trusted from './components/Trusted'
-import FeatureProducts from './components/FeatureProducts'
+import HeroSection from '../components/HeroSection'
+import Services from '../components/Services'
+import Trusted from '../components/Trusted'
+import FeatureProducts from '../components/FeatureProducts'
 
 const Wrapper = styled.section`
   background-color: ${({theme})=>theme.colors.bg};
